# Conflicts:
#	qtum wallet.xcodeproj/project.pbxproj
#	qtum wallet/TokenDetailsTableSource.h
#	qtum wallet/TokenDetailsViewController.m

+++ qtum wallet/Categories/NSData+Extension.m
@@ -3,7 +3,7 @@
 //  qtum wallet
 //
 //  Created by Vladimir Lebedevich on 18.04.17.
-//  Copyright © 2017 Designsters. All rights reserved.
+//  Copyright © 2017 PixelPlex. All rights reserved.
 //
 
 #import "NSData+Extension.h"
@@ -12,7 +12,7 @@
 
 + (NSData *)reverseData:(NSData *)data {
     const char *bytes = [data bytes];
-    int idx = [data length] - 1;
+    int idx = (int)([data length] - 1);
     char *reversedBytes = calloc(sizeof(char),[data length]);
     for (int i = 0; i < [data length]; i++) {
         reversedBytes[idx--] = bytes[i];
